# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Leedz API
  description: |
    API for Leedz - Sales Intelligence & CRM Platform
    
    ## Authentication
    All endpoints (except public ones) require a Bearer token in the Authorization header.
    
    ## Tenant Context
    The current tenant is determined from the JWT token. Use `/api/auth/switch-tenant` to change tenants.
    
    ## Rate Limiting
    - Default: 100 requests/minute
    - Search: 10 requests/minute
    - WhatsApp: 100 requests/hour
  version: 2.0.0
  contact:
    name: Leedz Team
    email: dev@leadz.sa
servers:
  - url: https://api.leadz.sa
    description: Production
  - url: https://api.staging.leadz.sa
    description: Staging
  - url: http://localhost:3000
    description: Development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input data
            details:
              type: object
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        emailVerified:
          type: boolean
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
    UserRole:
      type: string
      enum: [OWNER, ADMIN, MANAGER, SALES]
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            tenant:
              $ref: '#/components/schemas/Tenant'
            token:
              type: string
            expiresIn:
              type: integer
...
    JobType:
      type: string
      enum: [SEARCH, SURVEY, WHATSAPP, WHATSAPP_BULK, IMPORT, EXPORT, REVEAL, REPORT, SYNC]
    JobStatus:
      type: string
      enum: [PENDING, RUNNING, SUCCESS, FAILED, CANCELLED]
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/JobType'
        status:
          $ref: '#/components/schemas/JobStatus'
        progress:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
          nullable: true
        entityType:
          type: string
          nullable: true
        entityId:
          type: string
          format: uuid
          nullable: true
        result:
          type: object
          nullable: true
        error:
          type: object
          nullable: true
          properties:
            code:
              type: string
            message:
              type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true

    # ==================== Evidence & Reports ====================
...
    SendWhatsAppRequest:
      type: object
      required:
        - phone
        - message
      properties:
        leadId:
          type: string
          format: uuid
        phone:
          type: string
        message:
          type: string
        templateId:
          type: string
          format: uuid
      headers:
        X-Idempotency-Key:
          type: string
          description: Unique key to prevent duplicate sends

    # ==================== Team ====================
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login to existing account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/jobs:
    get:
      tags: [Jobs]
      summary: List jobs
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/JobStatus'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/JobType'
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      jobs:
                        type: array
                        items:
                          $ref: '#/components/schemas/Job'
  /api/jobs/{id}:
    get:
      tags: [Jobs]
      summary: Get job details
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      job:
                        $ref: '#/components/schemas/Job'

  # ==================== WhatsApp ====================
  /api/whatsapp/send:
    post:
      tags: [WhatsApp]
      summary: Send WhatsApp message
      security:
        - BearerAuth: []
      parameters:
        - name: X-Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendWhatsAppRequest'
      responses:
        '202':
          description: Message queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      jobId:
                        type: string
                        format: uuid
                      messageId:
                        type: string
                        format: uuid
...
  /api/agent/jobs/{jobId}/ack:
    post:
      tags: [Agent]
      summary: Acknowledge job receipt
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accepted:
                  type: boolean
                reason:
                  type: string
      responses:
        '200':
          description: Acknowledgment received
...